requirements.txt
asgiref==3.8.1
cx_Oracle==8.3.0
dj-database-url==2.3.0
Django==5.2.1
django-environ==0.12.0
django-jazzmin==3.0.1
django-jet-reboot==1.3.10
djangorestframework==3.16.0
dotenv==0.9.9
environ==1.0
google-api-python-client==1.4.1
httplib2==0.22.0
livereload==2.7.1
oauth2client==4.1.3
pillow==11.2.1
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dotenv==1.1.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tornado==6.5
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
whitenoise==6.9.0
